Initial set of Apache Flight improvements.
fix: compress the Arrow Schema using ZStandard when being written.

Requires adding an additional fields to the protocol buffer to store
the decompressed size of the schema.

fix: add caching of serialized schemas and compressed schemas

Since Arrow schemas don't often change once a process has started
allow the serialized schemas to be cached if there is a metadata key
set on the schema.

If the schema includes a metadata key "serialization_cache_key" use
that for the caching